added lib directory, improved the demos in the demos directory, added minesweeper and launchtest demos wrote a new colorpicker demo removed obsolete demos/extra directory and TTD file made demos in tndesigner directory work extended the CheckValidParent function in tnWidgets.c for menus and removed the fixme remark there got the Makefile in the microwindows/src directory to call the TinyWidgets Makefile to compile and clean silenced warnings in mwopenfile.c and mwterm renamed winlib/fileopen.c to winlib/openfile.c to make the name compatible with the WinAPI Openfiledialog.

+++ src/demos/nxscribble/sc.c
@@ -22,6 +22,7 @@ Software Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 ***********************************************************************/
 
 #include <string.h>
+#include <unistd.h>
 #include "bitvector.h"
 #include "matrix.h"
 #include "util.h"
@@ -409,7 +410,7 @@ register sClassifier sc;
 {
 	register sClassIndex c;
 
-	printf("\n----Classifier %x, %d features:-----\n", (int)sc, sc->nfeatures);
+	printf("\n----Classifier %x, %d features:-----\n", (intptr_t)sc, sc->nfeatures);
 	printf("%d classes: ", sc->nclasses);
 	for(c = 0; c < sc->nclasses; c++)
 		printf("%s  ", sc->classdope[c]->name);
